requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.8
pyyaml>=6.0
